Module-require
We now have the entire behavior of Menu available in order.js. Here, we notice:
In order.js we import the module by creating a variable with const called Menu and setting it equal to the value of the require() function. We can set this variable to any variable name we like, such as menuItems. 
require() is a JavaScript function that enables a module to load by passing in the module’s filepath as a parameter. 
'./menu.js' is the argument we pass to the require() function. 
The placeOrder() function then uses the Menu module in its function definition. By calling Menu.specialty, we access the property specialty defined in the Menu module. 
We can then invoke the function using placeOrder()
Taking a closer look, the pattern to import a module is:
Import the module 
Use the module and its properties within a program.