Create module_export.js
export default uses the JavaScript export statement to export JavaScript objects, functions, and primitive data types. 
Menu refers to the name of the Menu object, the object that we are setting the properties on within our modules.
When using ES6 syntax, we use export default in place of module.exports.